ARE-CARS-WASTING-SO-MUCH-FUEL-?
Project description
The target of the project is to analyze how different characteristics of an engine affects the fuel consumption of a car.#In order to do this a data set was downloaded, from kaggle, with characteristics of different cars.
Trying to understand which of these characteristics were more relevant.
Rstudio was used to analyze data.

 How to install and run the project
First of all, it is necessary to download the data set. It is a CSV file named Engines.
Then, download the Rstudio file called FinalProjectBigData.
Once this is done, open Rstudio file and change the location of the CSV file, to the place it is located now.
engine <- read.csv("writtingHereTheNewLocation/Engines.csv")
Finally, run the Rstudio code.